backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
pillow
pytesseract